framework: move into pkg/api
