+++ pkg/api/candidate_factory.go
@@ -1,4 +1,4 @@
-package framework
+package api
 
 import (
 	"math/rand"
