+++ pkg/api/evaluated_candidate_test.go
@@ -1,4 +1,4 @@
-package framework
+package api
 
 import (
 	"testing"
