+++ pkg/api/null_fitness_evaluator.go
@@ -1,4 +1,4 @@
-package framework
+package api
 
 // NullFitnessEvaluator is a fitness evaluator that always return a score of 0.
 //
@@ -8,12 +8,8 @@ type NullFitnessEvaluator struct{}
 
 // Fitness returns a score of zero, regardless of the candidate being
 // evaluated.
-func (fe NullFitnessEvaluator) Fitness(candidate Candidate, population []Candidate) float64 {
-	return 0
-}
+func (NullFitnessEvaluator) Fitness(candidate Candidate, population []Candidate) float64 { return 0 }
 
 // IsNatural always returns true. However, the return value of this method is
 // irrelevant since no meaningful fitness scores are produced.
-func (fe NullFitnessEvaluator) IsNatural() bool {
-	return true
-}
+func (NullFitnessEvaluator) IsNatural() bool { return true }
