+++ pkg/api/termination_condition.go
@@ -1,4 +1,4 @@
-package framework
+package api
 
 import "fmt"
